implement multi target in providing configuration
